query.cc: Cast time_t to unsigned long to avoid problems on 64bit platforms.
git-svn-id: file:///home/xapian-svn/xapian/trunk@6481 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-applications/omega/query.cc
@@ -1522,7 +1522,7 @@ eval(const string &fmt, const vector<string> &param)
 		break;
 	    case CMD_now: {
 		char buf[64];
-		sprintf(buf, "%u", time(NULL));
+		sprintf(buf, "%lu", (unsigned long)time(NULL));
 		value = buf;
 		break;
 	    }
